fix(code-scanning): address open alerts — overflow, XSS, format, vendored hljs (#1446)
Code scanning cleanup (alerts from the 07-27 CodeQL run):

- npu_engine_hybrid.cpp: cast int products to size_t/double at 38 flagged
  sites (vector sizes, buffer indexes, KV offsets) + float->double in
  RMS-norm accumulators (ks[d]*ks[d], qb[...]*qb[...], ...). No behavior
  change — prevents int overflow before conversion to size_t.
- backend_generic.cpp: (size_t) on tensor-size products in load_tensor
  (NH*HD*H, NKV*HD*H, H*NH*HD, FF*H, H*FF).
- zaya_engine.cpp: (size_t) on n_layers*rtr_h in ALLOC_OR_FAIL.
- chat.html: session ids via crypto.getRandomValues instead of
  Math.random (js/insecure-randomness); esc() now escapes quotes
  (js/incomplete-html-attribute-sanitization); highlight.js vendored
  locally, no CDN script (js/functionality-from-untrusted-source).
- test_pipeline.cpp: %zu for size_t layer_count (cpp/wrong-type-format-argument).

Stale alerts (deleted/renamed files: npu_engine_stdio.cpp, npu_engine_fused.cpp
-> .hip, npu-chatd.cpp, jarvis/*, npu_test_all_xclbins.cpp double-free, etc.)
auto-close on the next CodeQL run.

+++ src/zaya_engine.cpp
@@ -343,7 +343,7 @@ ZayaState* zaya_init(const char* weights_dir, const ZayaConfig* cfg) {
     ALLOC_OR_FAIL(s, alloc_f16, s->d_kout, eng.kd);
     ALLOC_OR_FAIL(s, alloc_f16, s->d_vout, eng.kd);
     ALLOC_OR_FAIL(s, alloc_f32, s->d_skip_flag, 1);
-    ALLOC_OR_FAIL(s, alloc_f32, s->d_prev_rs, eng.n_layers * eng.rtr_h);
+    ALLOC_OR_FAIL(s, alloc_f32, s->d_prev_rs, (size_t)eng.n_layers * eng.rtr_h);
     ALLOC_OR_FAIL(s, alloc_f32, s->d_expert_idx, 1);
     ALLOC_OR_FAIL(s, alloc_f32, s->d_expert_wt, 1);
     #undef ALLOC_OR_FAIL
